# docs/api/openapi/identity-service/openapi.yaml
openapi: 3.0.3
info:
  title: POS Identity Service API
  version: 0.1.0
  description: |
    OpenAPI documentation for `services/identity-service`.

    Notes on authentication:
    - The service issues JWT access tokens (Bearer) via `/auth/login` and `/auth/refresh`.
    - All protected endpoints require `Authorization: Bearer <JWT>`.
    - Admin/management endpoints are restricted to `ROLE_ADMIN`.
servers:
  - url: http://localhost:8081
tags:
  - name: Health
  - name: Auth
  - name: Users
  - name: Roles
  - name: Login Audit
security:
  - bearerAuth: []
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: OK Good

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain tokens
      security: []
      description: |
        Authenticates a user (must be `ACTIVE`) and returns an access token (JWT) and a refresh token.
      parameters:
        - in: header
          name: X-Forwarded-For
          required: false
          schema: { type: string }
          description: Forwarded client IP address (used for audit logging).
        - in: header
          name: User-Agent
          required: false
          schema: { type: string }
          description: Client user agent (used for audit logging).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "401":
          description: Invalid credentials or user not ACTIVE (empty body)

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "401":
          description: Refresh token not found/expired/revoked (empty body)

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: Logged out (always returns 204)

  /auth/login-audit:
    get:
      tags: [Login Audit]
      summary: List login audit logs
      parameters:
        - in: query
          name: username
          required: false
          schema: { type: string }
          description: Filter by username (exact match).
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoginAuditLogDto"

  /bootstrap/admin:
    post:
      tags: [Auth]
      summary: Break-glass provision first admin (disabled by default)
      description: |
        Creates the first `ADMIN` user in the system. This endpoint is **disabled by default** and must be enabled via
        `bootstrap.breakglass.enabled=true` in a controlled deployment.

        Required controls:
        - Provide a one-time bootstrap token via `X-Bootstrap-Token` (configured from secret/env).
        - Prefer mTLS / IP allowlisting at your gateway/ingress.
      security: []
      parameters:
        - in: header
          name: X-Bootstrap-Token
          required: true
          schema: { type: string }
        - in: header
          name: X-Forwarded-For
          required: false
          schema: { type: string }
        - in: header
          name: User-Agent
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBootstrapRequest"
      responses:
        "201":
          description: Admin created
        "400":
          description: Invalid request / password policy failure
        "403":
          description: Forbidden (bad token or denied by IP allowlist)
        "409":
          description: Admin already provisioned / username exists

  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: Relative URI of the created user.
              example: /users/123
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "500":
          description: Server error (e.g., role not found)

  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Users]
      summary: Get user by id
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "404":
          description: Not found (empty body)
    put:
      tags: [Users]
      summary: Update user (role/admin unit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "404":
          description: Not found (empty body)
        "500":
          description: Server error (e.g., role not found)
    delete:
      tags: [Users]
      summary: Delete user
      responses:
        "204":
          description: Deleted (even if id did not exist)

  /users/{id}/status:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    patch:
      tags: [Users]
      summary: Update user status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "404":
          description: Not found (empty body)

  /roles:
    get:
      tags: [Roles]
      summary: List roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleDto"
    post:
      tags: [Roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateRequest"
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: Relative URI of the created role.
              example: /roles/10
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleDto"

  /roles/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Roles]
      summary: Get role by id
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleDto"
        "404":
          description: Not found (empty body)
    put:
      tags: [Roles]
      summary: Update role (description)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleDto"
        "404":
          description: Not found (empty body)
    delete:
      tags: [Roles]
      summary: Delete role
      responses:
        "204":
          description: Deleted (even if id did not exist)

  /roles/{id}/permissions:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Roles]
      summary: List permissions assigned to a role
      responses:
        "200":
          description: Permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDto"
        "500":
          description: Server error (e.g., role not found)
    post:
      tags: [Roles]
      summary: Assign permission to a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionAssignRequest"
      responses:
        "200":
          description: Updated list of permissions for the role
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDto"
        "500":
          description: Server error (e.g., role not found)

  /roles/{roleId}/permissions/{permissionId}:
    parameters:
      - in: path
        name: roleId
        required: true
        schema: { type: integer, format: int64 }
      - in: path
        name: permissionId
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [Roles]
      summary: Remove permission from role
      responses:
        "204":
          description: Removed
        "500":
          description: Server error (e.g., permission not found or does not belong to role)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdminBootstrapRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        adminUnitId: { type: integer, format: int64, nullable: true }
    AuthRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    AuthTokensResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, tokenType]
      properties:
        accessToken: { type: string, description: JWT access token }
        refreshToken: { type: string, description: Opaque refresh token (UUID string) }
        tokenType: { type: string, example: Bearer }
    UserCreateRequest:
      type: object
      additionalProperties: false
      required: [username, password, roleId]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        roleId: { type: integer, format: int64 }
        adminUnitId: { type: integer, format: int64, nullable: true }
    UserUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        roleId: { type: integer, format: int64, nullable: true }
        adminUnitId: { type: integer, format: int64, nullable: true }
    UserStatusUpdateRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          description: User status string (e.g., ACTIVE, DISABLED).
    UserDto:
      type: object
      additionalProperties: false
      required: [id, username, status]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        roleName: { type: string, nullable: true }
        adminUnitId: { type: integer, format: int64, nullable: true }
        status: { type: string }
    RoleCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
    RoleUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        description: { type: string, nullable: true }
    RoleDto:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string, nullable: true }
    PermissionAssignRequest:
      type: object
      additionalProperties: false
      required: [permissionKey]
      properties:
        permissionKey: { type: string }
    PermissionDto:
      type: object
      additionalProperties: false
      required: [id, permissionKey]
      properties:
        id: { type: integer, format: int64 }
        permissionKey: { type: string }
    LoginAuditLogDto:
      type: object
      additionalProperties: false
      required: [id, username, success, createdAt]
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64, nullable: true }
        username: { type: string }
        success: { type: boolean }
        ipAddress: { type: string, nullable: true }
        userAgent: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }