# services/identity-service/src/main/java/com/pos/identity_service/dto/AuthTokensResponse.java
package com.pos.identity_service.dto;

public class AuthTokensResponse {

    private String accessToken;
    private String refreshToken;
    private String tokenType = "Bearer";

    public AuthTokensResponse() {
    }

    public AuthTokensResponse(String accessToken, String refreshToken) {
        this.accessToken = accessToken;
        this.refreshToken = refreshToken;
    }

    public String getAccessToken() {
        return accessToken;
    }

    public void setAccessToken(String accessToken) {
        this.accessToken = accessToken;
    }

    public String getRefreshToken() {
        return refreshToken;
    }

    public void setRefreshToken(String refreshToken) {
        this.refreshToken = refreshToken;
    }

    public String getTokenType() {
        return tokenType;
    }

    public void setTokenType(String tokenType) {
        this.tokenType = tokenType;
    }
}