# services/identity-service/src/main/java/com/pos/identity_service/dto/RefreshTokenRequest.java
package com.pos.identity_service.dto;

public class RefreshTokenRequest {

    private String refreshToken;

    public String getRefreshToken() {
        return refreshToken;
    }

    public void setRefreshToken(String refreshToken) {
        this.refreshToken = refreshToken;
    }
}