# services/identity-service/src/main/java/com/pos/identity_service/dto/RoleUpdateRequest.java
package com.pos.identity_service.dto;

public class RoleUpdateRequest {

    private String description;

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}